+++ src/xpn_client/xpn/xpn_simple/xpncore/xpn_f.c
@@ -407,15 +407,16 @@ int xpn_fseek(FILE *stream, long offset, int whence)
 	//FIXME: Needs lots of debugging!
 	int off;
 
-   switch(whence){
+   switch(whence)
+   {
 	case SEEK_SET:
-		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_SET)\n", stream->_fileno, offset);
+		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_SET)\n", stream->_fileno, (unsigned long)offset);
 		break;
 	case SEEK_CUR:
-		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_CUR)\n", stream->_fileno, offset);
+		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_CUR)\n", stream->_fileno, (unsigned long)offset);
 		break;
 	case SEEK_END:
-		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_END)\n", stream->_fileno, offset);
+		fprintf(stderr, "Calling xpn_fseek(%d, %lu, SEEK_END)\n", stream->_fileno, (unsigned long)offset);
 		break;
 	default:
 		fprintf(stderr, "Calling xpn_fseek(%d, %lu, %d)\n", stream->_fileno, (unsigned long)offset, whence);
